revert: reverted timeout on Worker
The issue is that on big wallet, it might takes more than that. This cannot be implemented this way. We need a proper implementation that takes account on potential activity of the Worker. Keeping in mind that sometimes, activity might happen while we still stalled on a action and didn't release any STARTED event